Add DriverSettings class
It's not very convenient to do 2 ioctls in the same place,
because some drivers may not support some ioctls. For example
virtio driver can't return interface settings, but ioctl, which
returns driver settings works. So to have ability to get
driver info let's add second class, DriverSettings and use it
in InterfaceSettings (for compatibility).